Fix async. IAM credential fetching and debugging not working w/ IAM
So apparently statsd will change the `metrics` object on the next
event loop run. This causes issues since this backend fetches the
IAM credentials from the EC2 meta data.

During that async IAM fetch, statsd changes `metrics` and by the time
we are ready to submit to cloudwatch, all the values have been reset.

Added in a quick / dirty object cloning on metrics so we can avoid
this side effect.